doc: remove pull request material that is for the wrong audience
The pull request document is long, which is fine, but only if it is all
relevant to someone opening a pull request. Including lengthy material
geared toward a different audience is disrespectful to the contributor.
The more material we provide people to wade through, the less likely
they are to read things carefully, and therefore the less likely they
are to understand the other parts that are actually for them and
therefore more important.

This section being removed is geared towards collaborators. Much, if not
all, of the material is replicated elsewhere where collaborators would
see it. Some of the material here is even repeated in this very document
(e.g., the part about CI requirements). Some of it is out of date (e.g.,
the part about only collaborators being able to start CI--triagers can
do that too) which is almost inevitable when information is duplicated
within the repo like this.

This document is linked in the GitHub interface when someone is trying
to open a pull request. It should be geared to the person opening a pull
request in the hope that they will read and recall the information.